Adding a default BatchConfiguration and refactoring BatchOverrideConfiguration into a batchmanager package
