+++ core/sdk-core/src/main/java/software/amazon/awssdk/core/batchmanager/BatchOverrideConfiguration.java
@@ -13,19 +13,19 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.core;
+package software.amazon.awssdk.core.batchmanager;
 
 import java.time.Duration;
-import java.util.concurrent.ScheduledExecutorService;
+import java.util.Optional;
 import software.amazon.awssdk.annotations.SdkPublicApi;
-import software.amazon.awssdk.core.internal.batchmanager.BatchManager;
+import software.amazon.awssdk.core.internal.batchmanager.DefaultBatchManager;
 import software.amazon.awssdk.utils.ToString;
 import software.amazon.awssdk.utils.Validate;
 import software.amazon.awssdk.utils.builder.CopyableBuilder;
 import software.amazon.awssdk.utils.builder.ToCopyableBuilder;
 
 /**
- * Configuration values for the {@link BatchManager}. All values are optional, and the default values will be used
+ * Configuration values for the {@link DefaultBatchManager}. All values are optional, and the default values will be used
  * if they are not specified.
  */
 @SdkPublicApi
@@ -34,14 +34,10 @@ public final class BatchOverrideConfiguration implements ToCopyableBuilder<Batch
 
     private final Integer maxBatchItems;
     private final Duration maxBatchOpenInMs;
-    private final ScheduledExecutorService scheduledExecutor;
 
     public BatchOverrideConfiguration(Builder builder) {
-        Validate.notNull(builder.maxBatchItems, "maxBatchItems cannot be null");
-        this.maxBatchItems = Validate.isPositive(builder.maxBatchItems, "maxBatchItems");
-        Validate.notNull(builder.maxBatchOpenInMs, "maxBatchOpenInMs cannot be null");
-        this.maxBatchOpenInMs = Validate.isPositive(builder.maxBatchOpenInMs, "maxBachOpenInMs");
-        this.scheduledExecutor = Validate.notNull(builder.scheduledExecutor, "scheduledExecutor cannot be null");
+        this.maxBatchItems = Validate.isPositiveOrNull(builder.maxBatchItems, "maxBatchItems");
+        this.maxBatchOpenInMs = Validate.isPositiveOrNull(builder.maxBatchOpenInMs, "maxBachOpenInMs");
     }
 
     public static Builder builder() {
@@ -51,27 +47,22 @@ public final class BatchOverrideConfiguration implements ToCopyableBuilder<Batch
     /**
      * @return the optional maximum number of messages that are batched together in a single request.
      */
-    public Integer maxBatchItems() {
-        return maxBatchItems;
+    public Optional<Integer> maxBatchItems() {
+        return Optional.ofNullable(maxBatchItems);
     }
 
     /**
      * @return the optional maximum amount of time (in milliseconds) that an outgoing call waits to be batched with messages of
      * the same type.
      */
-    public Duration maxBatchOpenInMs() {
-        return maxBatchOpenInMs;
-    }
-
-    public ScheduledExecutorService scheduledExecutor() {
-        return scheduledExecutor;
+    public Optional<Duration> maxBatchOpenInMs() {
+        return Optional.ofNullable(maxBatchOpenInMs);
     }
 
     @Override
     public Builder toBuilder() {
         return new Builder().maxBatchItems(maxBatchItems)
-                            .maxBatchOpenInMs(maxBatchOpenInMs)
-                            .scheduledExecutor(scheduledExecutor);
+                            .maxBatchOpenInMs(maxBatchOpenInMs);
     }
 
     @Override
@@ -96,17 +87,13 @@ public final class BatchOverrideConfiguration implements ToCopyableBuilder<Batch
         if (maxBatchItems != null ? !maxBatchItems.equals(that.maxBatchItems) : that.maxBatchItems != null) {
             return false;
         }
-        if (maxBatchOpenInMs != null ? !maxBatchOpenInMs.equals(that.maxBatchOpenInMs) : that.maxBatchOpenInMs != null) {
-            return false;
-        }
-        return scheduledExecutor.equals(that.scheduledExecutor);
+        return maxBatchOpenInMs != null ? maxBatchOpenInMs.equals(that.maxBatchOpenInMs) : that.maxBatchOpenInMs == null;
     }
 
     @Override
     public int hashCode() {
         int result = maxBatchItems != null ? maxBatchItems.hashCode() : 0;
         result = 31 * result + (maxBatchOpenInMs != null ? maxBatchOpenInMs.hashCode() : 0);
-        result = 31 * result + scheduledExecutor.hashCode();
         return result;
     }
 
@@ -114,7 +101,6 @@ public final class BatchOverrideConfiguration implements ToCopyableBuilder<Batch
 
         private Integer maxBatchItems;
         private Duration maxBatchOpenInMs;
-        private ScheduledExecutorService scheduledExecutor;
 
         private Builder() {
         }
@@ -126,11 +112,6 @@ public final class BatchOverrideConfiguration implements ToCopyableBuilder<Batch
 
         public Builder maxBatchOpenInMs(Duration maxBatchOpenInMs) {
             this.maxBatchOpenInMs = maxBatchOpenInMs;
-            return this;
-        }
-
-        public Builder scheduledExecutor(ScheduledExecutorService scheduledExecutor) {
-            this.scheduledExecutor = scheduledExecutor;
             return this;
         }
